CSV upload
Here is a dummy file that i got a csv upload to work.
I created an excel spreadsheet with admin, admin, admin, /n user, password, role and saved it as CSV file. 
This '/upload' erb successfully handled the upload of the file. We just need to send that data to the database, salt, and hash